[disk] `use_df` parameter to force use `df`
`psutil.disk_usage` method uses the `statvfs` POSIX API to resolve the
disks' usage. **Problem?** This call can hang "forever" when a file
system is listed but not available. Worse, it can’t get interrupted or
killed, making the Watchdog.

I attached more references to this problem below.

> Basically if you have a network filesystem listed in /proc/mounts and
  it is unreachable (e.g. because there is no network), statvfs will
  hang on stat’ing the network directory, even if you called statvfs on
  a completely different directory

To workaround this issue, add an optional `use_df` parameter to instruct
the check to use the `df` command instead of the `psutil` module. `df`
uses the `statvfs` POSIX API too, but prints an appropriate error message
when it fails instead of hanging.

**References**
https://sourceware.org/ml/libc-hacker/2003-09/msg00056.html
http://ilostmynotes.blogspot.com/2014/05/cross-platform-python-method-for.html
https://www.redhat.com/archives/rhl-list/2004-March/msg01647.html